server/pyproject.toml
[tool.poetry]
name = "cshr"
version = "0.1.0"
description = ""
authors = ["Mahmoud-Emad <mahmmoud.hassanein@gmail.com>"]



[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
drf-yasg = "^1.21.3"
djangorestframework = "^3.13.1"
django-debug-toolbar = "^3.5.0"
djangorestframework-simplejwt = "^5.2.0"
Pillow = "^9.2.0"
django-cors-headers = "^3.13.0"
tzdata = "^2022.2"
django-celery-beat = "^2.3.0"
python-decouple = "^3.6"
redis = "^4.3.4"
whitenoise = "^6.2.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"